fix concurrent map write
